Error on unfound resource specification
There are cases where we are looking for a specified resource and can't
find it due to a configuration error.  We didn't supply a hint on what
is missing, rather yielding nil.

This changes that with at least showing an error message (in a
backtrace) that hints at which entity was not found.