Thread safety enhancements
Some hash-tables which are read-only did not have :synchronized t set to
them.  This has performance benefits but behaviour is undefined with
respect to threading so it *might* create issues.  We are setting these
as :synchronized t for now though we should work towards having this
information available thread-local instead because the current
implementation will create many unnecessary locks.