Fetching multiple items's properties in one go with construct
A CONSTRUCT is well suited for fetching the data from multiple items in
one go.  This approach lets us get away with less queries.  Some
splitting is possible for the future.  This approach should be less
heavy on the database because there are less possible issues with
orthogonal requests exploding the solution space.