Correctly parse typed literal values
The current construct does not use :typed-literal as a type (which it
seems to have done incorrectly before).  This results in the parsing not
happening when it should happen.  Looking at the spec it should be
:literal, and looking at our implemenation it should be :typed-literal
in these cases.  We have updated our code to cope with this case
regardless of what is yielded by the sparql endpoint.