Parse more format variants
It seems CONSTRUCT query responses are sent differently than the regular
query responses.  We now support different approaches for parsing such
content and assume this will work with the combined approach we receive
today.  Ideally this approach is altered for something more robust in
the distant future but this will do for as far as we can see today.