Optionally filter types in CONSTRUCT using FILTER instead
This lets us fetch all types whilst still ensuring the necessary types
are part of the solution.  This approach may be heavier on the system
that creates the solutions but it could be sufficient for constructing
answers in general.  We hope this solution could support some buggy
systems and we hope that further work on this feature may help us detect
inconsistencies in datasets.