Add back support for empty included relations
When an `include=` is supplied which has no relations, the data
attribute would not be set.  This commit corrects that.

The system uses the availability of a (possibly empty) hash value having
been set as an indication that a relationship should be included on the
resource.  This relationship was only initialized for cases where we
could find related resources through the CONSTRUCT query.  When no data
exists we should explicitly set the relation to be empty.  Which is what
the updated code snipped does.