Optionally use native booleans rather than custom booleans
mu-cl-resources ships with a custom boolean implementation because our
triplestore of choice had strange issues handling xsd:boolean.  This
change adds a configuration parameter
`mu-support:*use-custom-boolean-type-p*` which wen set to `nil` will
fall back to writing xsd:boolean instead.

Note that this only influences the writing of booleans.  Reading of
booleans is supported in both formats.  Writing data which is in one of
these formats (A) with a mu-cl-resources which writes the other
format (B), will remove the value of format A and set it in format B.
Data can thus gradually be changed into the newly preferred format.