Flush sparql updates before sending queries, allow out of scope queries
Previous implementation barfed at queries being sent whilst within a
query-update-group (a series of queries that can be sent in one go).
The current implementation flushes the pending update queries before
sending the select query which seems to be the safest bet.

Furthermore a macro is provided which allows to execute queries out of
the scope of a pending update group for cases where you know the data
exists in the triplestore.

co-author: @erikap