Retry fetching classes on external calls when necessary
Some cases seem to show classes not being found where they should be.
This may be due to a cache optimization issue or a locking issue with
partial requests being handled by the caches.  This makes the system
retry to fetch the types on such cases which should resolve some
stability issues when hammering resources.