Better stacktrace enhanced type retry
The stack trace now contains the full unwound stack trace, rather than
the one where we detect the error.

Not finding the type of a thing now happens in two passes with a brief
pause in between.  This pause may help incomplete data to be fully
pushed through at the cost of a longer response time.