Nicer warning message for single valued property with multiple values
When a single valued property has multiple values we emit a warning message.
The message itself is well received.  The content of the message can be made
nicer to read.  Here it is presented as a list with dashes.  As long as messages
don't overlap, this should be nicer to read.