Approach for allowing unknown literal data types
It may be that the sparql endpoint contains unknown literal datatypes.  It's not
obvious what we want to do in that case. We could read them as a string but
that's likely not the very best approach.  Still, it may be the only way out for
stacks in which we can't control the literal's datatype.  We now yield a more
informative error.  The error is still a 'simple-error but could become a more
fine-grained error object in the future.