Extract and cache calculation of incoming relations to delete
When deleting relations, the resources which point to us but to whom we do not
point, also need to be removed from the triplestore.  This code tackles that
case.  Whether this is an absolute must or not is up for debate but it will not
cause extra surprises and may cover some edge cases.  Caching helps to limit the
performance impact.