Process delta messages based on an in-memory queue
This approach should leave more connections open for effectively answering
questions of end-users, but it may cause more memory-consumption than we'd
ideally like.

Current implementation also keeps the system awake which is not desired when
working on battery power.  This can be improved by monitoring the lock itself.

This iteration is sufficient to verify memory use is reasonably kept in check.